# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Risk Model API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LoanApplicationHomeOwnership =
  (typeof LoanApplicationHomeOwnership)[keyof typeof LoanApplicationHomeOwnership];

export const LoanApplicationHomeOwnership = {
  rent: "rent",
  own: "own",
  mortgage: "mortgage",
} as const;

export type LoanApplicationLoanPurpose =
  (typeof LoanApplicationLoanPurpose)[keyof typeof LoanApplicationLoanPurpose];

export const LoanApplicationLoanPurpose = {
  debt_consolidation: "debt_consolidation",
  home_improvement: "home_improvement",
  major_purchase: "major_purchase",
  medical: "medical",
  business: "business",
  other: "other",
} as const;

export interface LoanApplication {
  applicantName: string;
  /** Age in years */
  age: number;
  /** Annual income in USD */
  income: number;
  /** Requested loan amount in USD */
  loanAmount: number;
  /** Loan term in months */
  loanTerm: number;
  /** Credit score (300-850) */
  creditScore: number;
  /** Existing debt to income ratio (0-1) */
  debtToIncomeRatio: number;
  /** Years at current employer */
  employmentYears: number;
  /** Number of open credit lines */
  numCreditLines: number;
  /** Number of late payments in last 2 years */
  numLatePayments: number;
  homeOwnership: LoanApplicationHomeOwnership;
  loanPurpose: LoanApplicationLoanPurpose;
}

export type CreditRiskResultRiskCategory =
  (typeof CreditRiskResultRiskCategory)[keyof typeof CreditRiskResultRiskCategory];

export const CreditRiskResultRiskCategory = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type CreditRiskResultDecision =
  (typeof CreditRiskResultDecision)[keyof typeof CreditRiskResultDecision];

export const CreditRiskResultDecision = {
  approved: "approved",
  review: "review",
  declined: "declined",
} as const;

export interface FeatureContribution {
  feature: string;
  value: number;
  /** Positive means increases risk, negative means decreases risk */
  contribution: number;
  displayName: string;
}

export interface CreditRiskResult {
  /** Probability of default (0-1) */
  defaultProbability: number;
  riskCategory: CreditRiskResultRiskCategory;
  /** Risk score (0-100, lower is better) */
  riskScore: number;
  decision: CreditRiskResultDecision;
  /** AI-generated plain-language explanation */
  aiExplanation: string;
  featureContributions: FeatureContribution[];
  id: number;
}

export interface CreditRiskRecord {
  id: number;
  applicantName: string;
  creditScore: number;
  income: number;
  loanAmount: number;
  defaultProbability: number;
  riskCategory: string;
  decision: string;
  createdAt: string;
}

export interface FeatureInfo {
  name: string;
  displayName: string;
  importance: number;
  description: string;
}

export interface ModelInfo {
  modelType: string;
  accuracy: number;
  auc: number;
  features: FeatureInfo[];
  trainingSamples: number;
  lastTrained: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}