# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Risk Model API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Runs logistic regression model on input features and returns a risk score with AI explanation
 * @summary Predict loan default probability
 */
export const PredictCreditRiskBody = zod.object({
  applicantName: zod.string(),
  age: zod.number().describe("Age in years"),
  income: zod.number().describe("Annual income in USD"),
  loanAmount: zod.number().describe("Requested loan amount in USD"),
  loanTerm: zod.number().describe("Loan term in months"),
  creditScore: zod.number().describe("Credit score (300-850)"),
  debtToIncomeRatio: zod
    .number()
    .describe("Existing debt to income ratio (0-1)"),
  employmentYears: zod.number().describe("Years at current employer"),
  numCreditLines: zod.number().describe("Number of open credit lines"),
  numLatePayments: zod
    .number()
    .describe("Number of late payments in last 2 years"),
  homeOwnership: zod.enum(["rent", "own", "mortgage"]),
  loanPurpose: zod.enum([
    "debt_consolidation",
    "home_improvement",
    "major_purchase",
    "medical",
    "business",
    "other",
  ]),
});

export const PredictCreditRiskResponse = zod.object({
  defaultProbability: zod.number().describe("Probability of default (0-1)"),
  riskCategory: zod.enum(["low", "medium", "high"]),
  riskScore: zod.number().describe("Risk score (0-100, lower is better)"),
  decision: zod.enum(["approved", "review", "declined"]),
  aiExplanation: zod
    .string()
    .describe("AI-generated plain-language explanation"),
  featureContributions: zod.array(
    zod.object({
      feature: zod.string(),
      value: zod.number(),
      contribution: zod
        .number()
        .describe(
          "Positive means increases risk, negative means decreases risk",
        ),
      displayName: zod.string(),
    }),
  ),
  id: zod.number(),
});

/**
 * @summary Get prediction history
 */
export const GetCreditRiskHistoryResponseItem = zod.object({
  id: zod.number(),
  applicantName: zod.string(),
  creditScore: zod.number(),
  income: zod.number(),
  loanAmount: zod.number(),
  defaultProbability: zod.number(),
  riskCategory: zod.string(),
  decision: zod.string(),
  createdAt: zod.date(),
});
export const GetCreditRiskHistoryResponse = zod.array(
  GetCreditRiskHistoryResponseItem,
);

/**
 * @summary Get model information and feature importances
 */
export const GetModelInfoResponse = zod.object({
  modelType: zod.string(),
  accuracy: zod.number(),
  auc: zod.number(),
  features: zod.array(
    zod.object({
      name: zod.string(),
      displayName: zod.string(),
      importance: zod.number(),
      description: zod.string(),
    }),
  ),
  trainingSamples: zod.number(),
  lastTrained: zod.string(),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});