# lib/api-zod/src/generated/types/creditRiskRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Risk Model API
 * OpenAPI spec version: 0.1.0
 */

export interface CreditRiskRecord {
  id: number;
  applicantName: string;
  creditScore: number;
  income: number;
  loanAmount: number;
  defaultProbability: number;
  riskCategory: string;
  decision: string;
  createdAt: Date;
}