# lib/api-zod/src/generated/types/creditRiskResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Risk Model API
 * OpenAPI spec version: 0.1.0
 */
import type { CreditRiskResultDecision } from "./creditRiskResultDecision";
import type { CreditRiskResultRiskCategory } from "./creditRiskResultRiskCategory";
import type { FeatureContribution } from "./featureContribution";

export interface CreditRiskResult {
  /** Probability of default (0-1) */
  defaultProbability: number;
  riskCategory: CreditRiskResultRiskCategory;
  /** Risk score (0-100, lower is better) */
  riskScore: number;
  decision: CreditRiskResultDecision;
  /** AI-generated plain-language explanation */
  aiExplanation: string;
  featureContributions: FeatureContribution[];
  id: number;
}