# lib/api-zod/src/generated/types/featureContribution.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Risk Model API
 * OpenAPI spec version: 0.1.0
 */

export interface FeatureContribution {
  feature: string;
  value: number;
  /** Positive means increases risk, negative means decreases risk */
  contribution: number;
  displayName: string;
}