# lib/api-zod/src/generated/types/featureInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Risk Model API
 * OpenAPI spec version: 0.1.0
 */

export interface FeatureInfo {
  name: string;
  displayName: string;
  importance: number;
  description: string;
}