# lib/api-zod/src/generated/types/loanApplication.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Risk Model API
 * OpenAPI spec version: 0.1.0
 */
import type { LoanApplicationHomeOwnership } from "./loanApplicationHomeOwnership";
import type { LoanApplicationLoanPurpose } from "./loanApplicationLoanPurpose";

export interface LoanApplication {
  applicantName: string;
  /** Age in years */
  age: number;
  /** Annual income in USD */
  income: number;
  /** Requested loan amount in USD */
  loanAmount: number;
  /** Loan term in months */
  loanTerm: number;
  /** Credit score (300-850) */
  creditScore: number;
  /** Existing debt to income ratio (0-1) */
  debtToIncomeRatio: number;
  /** Years at current employer */
  employmentYears: number;
  /** Number of open credit lines */
  numCreditLines: number;
  /** Number of late payments in last 2 years */
  numLatePayments: number;
  homeOwnership: LoanApplicationHomeOwnership;
  loanPurpose: LoanApplicationLoanPurpose;
}