# lib/api-zod/src/generated/types/modelInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Risk Model API
 * OpenAPI spec version: 0.1.0
 */
import type { FeatureInfo } from "./featureInfo";

export interface ModelInfo {
  modelType: string;
  accuracy: number;
  auc: number;
  features: FeatureInfo[];
  trainingSamples: number;
  lastTrained: string;
}